% Auto-generated: do not edit by hand
\name{dashWtgviewer}

\alias{dashWtgviewer}

\title{DashWtgviewer component}

\description{

}

\usage{
dashWtgviewer(id=NULL, environment=NULL, map=NULL, model=NULL,
results=NULL, show_map=NULL, stats=NULL, tooltip=NULL)
}

\arguments{
\item{id}{Character. }

\item{environment}{Logical. }

\item{map}{}

\item{model}{}

\item{results}{}

\item{show_map}{Logical. }

\item{stats}{Logical. }

\item{tooltip}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

