% Auto-generated: do not edit by hand
\name{dashWtgviewer}

\alias{dashWtgviewer}

\title{DashWtgviewer component}

\description{

}

\usage{
dashWtgviewer(id=NULL, colorscale=NULL, colorscale_clicked=NULL,
environment=NULL, map=NULL, model=NULL, results=NULL,
show_map=NULL, stats=NULL, tooltip=NULL)
}

\arguments{
\item{id}{Character. }

\item{colorscale}{Lists containing elements 'visible', 'min', 'max'.
those elements have the following types:
  - visible (logical; optional)
  - min (numeric; optional)
  - max (numeric; optional). }

\item{colorscale_clicked}{Logical. }

\item{environment}{Logical. }

\item{map}{}

\item{model}{}

\item{results}{}

\item{show_map}{Logical. }

\item{stats}{Logical. }

\item{tooltip}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

