% Auto-generated: do not edit by hand
\name{dashWtgviewer}

\alias{dashWtgviewer}

\title{DashWtgviewer component}

\description{

}

\usage{
dashWtgviewer(id=NULL, map=NULL, model=NULL, sea=NULL, tooltip=NULL)
}

\arguments{
\item{id}{Character. }

\item{map}{}

\item{model}{}

\item{sea}{Logical. }

\item{tooltip}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

